DocumentClassify/requirements.txt
streamlit
standard-imghdr
altair==4.2.2
pytesseract
tesseract
Pillow
spacy
transformers
en-core-web-sm-abd
python-docx
PyMuPDF
boto3
pdfminer
pandas
openpyxl
xlrd
google-generativeai


